// src/types/products-amnesia.ts
// Generated by https://quicktype.io
// endpoint:
// https://cms.aipus.co/api/products?populate=media&pagination[page]=1&pagination[pageSize]=9&filter[subcategories]=4

export interface ProductsAmnesia {
    data: ProductsAmnesiaDatum[];
    meta: Meta;
}

export interface ProductsAmnesiaDatum {
    id:         number;
    attributes: PurpleAttributes;
}

export interface PurpleAttributes {
    id_product:          string;
    uuid:                string;
    name:                string;
    description:         string;
    color:               string;
    width:               string;
    height:              string;
    brand:               string;
    stock:               string;
    base_price:          string;
    sale_price:          string;
    discount_price:      string;
    wholesaler_price:    string;
    links_marketplace:   LinksMarketplace;
    links_socialmedia:   LinksSocialmedia | null;
    sold:                string;
    state_product:       boolean;
    state_wholesaler:    boolean;
    state_discount:      boolean;
    deleted:             boolean;
    created:             string;
    updated:             string;
    createdAt:           string;
    updatedAt:           string;
    publishedAt:         string;
    locale:              string;
    quantity_wholesaler: null;
    other_data:          OtherData | null;
    media:               Media;
}

export interface LinksMarketplace {
    mercadolibre:     string;
    mercadoshops:     string;
    pay_mercadoshops: string;
    whatsapp:         string;
}

export interface LinksSocialmedia {
    nombre:      string;
    description: string;
}

export interface Media {
    data: MediaDatum[];
}

export interface MediaDatum {
    id:         number;
    attributes: FluffyAttributes;
}

export interface FluffyAttributes {
    name:              string;
    alternativeText:   null;
    caption:           null;
    width:             number;
    height:            number;
    formats:           Formats;
    hash:              string;
    ext:               EXT;
    mime:              MIME;
    size:              number;
    url:               string;
    previewUrl:        null;
    provider:          Provider;
    provider_metadata: null;
    createdAt:         string;
    updatedAt:         string;
}

export enum EXT {
    Webp = ".webp",
}

export interface Formats {
    thumbnail: Large;
    large?:    Large;
    medium?:   Large;
    small?:    Large;
}

export interface Large {
    name:   string;
    hash:   string;
    ext:    EXT;
    mime:   MIME;
    path:   null;
    width:  number;
    height: number;
    size:   number;
    url:    string;
}

export enum MIME {
    ImageWebp = "image/webp",
}

export enum Provider {
    Local = "local",
}

export interface OtherData {
}

export interface Meta {
    pagination: Pagination;
}

export interface Pagination {
    page:      number;
    pageSize:  number;
    pageCount: number;
    total:     number;
}